+++ cvescan/output_formatters/cli_output_formatter.py
@@ -1,4 +1,5 @@
 from sys import stdout
+from textwrap import wrap
 from typing import List
 
 from tabulate import tabulate
@@ -8,6 +9,7 @@ from cvescan import TargetSysInfo
 from cvescan.output_formatters import (
     AbstractOutputFormatter,
     AbstractStackableScanResultSorter,
+    ScanStats,
 )
 from cvescan.scan_result import ScanResult
 
@@ -38,9 +40,20 @@ class CLIOutputFormatter(AbstractOutputFormatter):
         priority_results = self._filter_on_priority(scan_results)
         fixable_results = self._filter_on_fixable(priority_results)
 
+        stats = self._get_scan_stats(scan_results, sysinfo)
+
+        msg = ""
         table_msg = self._format_table(priority_results, fixable_results, sysinfo)
-        summary_msg = self._format_summary(scan_results, sysinfo)
-        msg = "\n%s\n\n%s" % (table_msg, summary_msg)
+
+        if table_msg:
+            msg += "%s\n\n" % table_msg
+
+        summary_msg = self._format_summary(stats, sysinfo)
+        msg += summary_msg
+
+        suggestions_msg = self._format_suggestions(stats, sysinfo)
+        if suggestions_msg:
+            msg += "\n\n%s" % suggestions_msg
 
         return_code = CLIOutputFormatter._determine_return_code(
             priority_results, fixable_results
@@ -48,9 +61,7 @@ class CLIOutputFormatter(AbstractOutputFormatter):
 
         return (msg, return_code)
 
-    def _format_summary(self, scan_results: List[ScanResult], sysinfo: TargetSysInfo):
-        stats = self._get_scan_stats(scan_results, sysinfo)
-
+    def _format_summary(self, stats: ScanStats, sysinfo: TargetSysInfo):
         # Disabling for now
         # apps_enabled =
         # CLIOutputFormatter._format_esm_enabled(sysinfo.esm_apps_enabled)
@@ -207,3 +218,23 @@ class CLIOutputFormatter(AbstractOutputFormatter):
             return str(value)
 
         return "\u001b[38;5;%dm%s\u001b[0m" % (color_code, str(value))
+
+    def _format_suggestions(self, stats: ScanStats, sysinfo: TargetSysInfo):
+        ua_msg = (
+            "%d additional security patch(es) are available if Ubuntu Advantage for %s "
+            "is enabled. For more information, see %s."
+        )
+
+        if stats.infra_vulns > 0 and sysinfo.esm_infra_enabled == False:  # noqa: E712
+            infra_msg = ua_msg % (
+                stats.infra_vulns,
+                "Infrastructure",
+                const.UA_INFRA_URL,
+            )
+            return CLIOutputFormatter._wrap_text(infra_msg)
+
+        return ""
+
+    @staticmethod
+    def _wrap_text(text):
+        return "\n".join(wrap(text, 88))